Extract Cell Run Toolbar.
